replace config to Makefile

+++ Makefile
@@ -22,12 +22,23 @@ build_downloader:
 		--build-arg UID=$(UID) \
 		./repo_downloader || true
 
-clone: build_downloader
+config:
+	echo "BITBUCKET_USERNAME=$(NICKNAME)\n\
+	BITBUCKET_APP_PASSWORD=$(APPPASS)\n\
+	BITBUCKET_TEAM_NAME=hexlet\n\
+	BITBUCKET_API_URL=https://api.bitbucket.org/2.0/repositories/" > bitbucket.config.env
+
+clone:
+ifneq ($(shell test -f bitbucket.config.env && echo true), true)
+	@ echo "Please, run 'make config NICKNAME=username APPPASS=paSsw0rd', see README for details" >&2; exit 1;
+else
+	make build_downloader
 	docker run --rm -it --name hexlet_downloader \
 		-v $(SSH_KEYS_PATH):/home/$(UNAME)/.ssh \
 		-v $(CURDIR):/repos \
 		--env-file ./bitbucket.config.env \
 		$(DOWNLOADER_IMAGE_NAME):latest $(DOWNLOADER_FLAG)
+endif
 
-rebase: build_downloader
+rebase:
 	make clone DOWNLOADER_FLAG=--update
